Fixing SA1019 deprecated
SA1019: aws.EndpointResolverWithOptionsFunc is deprecated: The global
endpoint resolution interface is deprecated. See deprecation docs on
[EndpointResolver]. (staticcheck)

SA1019: aws.Endpoint is deprecated: This structure was used with the
global [EndpointResolver] interface, which has been deprecated in favor
of service-specific endpoint resolution. See the deprecation docs on
that interface for more information. (staticcheck)

SA1019: aws.Endpoint is deprecated: This structure was used with the
global [EndpointResolver] interface, which has been deprecated in favor
of service-specific endpoint resolution. See the deprecation docs on
that interface for more information. (staticcheck)

SA1019: aws.Endpoint is deprecated: This structure was used with the global [EndpointResolver] interface, which has been deprecated in favor of service-specific endpoint resolution. See the deprecation docs on that interface for more information. (staticcheck)